# apps/api/src/provider-protocol-routes.mjs
import { validateTextProfile } from '../../../src/provider-config/text-profile.mjs';
import { protocolDigest } from '../../../src/provider-config/protocol-confirmations.mjs';

const fail = (key, statusCode = 422) => { throw Object.assign(new Error(key), { statusCode, errorKey: key }); };
const body = (request, allowed) => { const value = request.body; if (!value || typeof value !== 'object' || Array.isArray(value) || Object.keys(value).some((key) => !allowed.includes(key))) fail('invalid_request'); return value; };
const stateDigest = ({ protocolId, version, baseVersion, state }) => protocolDigest({ protocolId, version, baseVersion, state });
const session = (auth) => { if (auth.authMethod !== 'session' || !auth.sessionId) fail('step_up_required', 403); return auth.sessionId; };

export function registerProviderProtocolRoutes(app, { directory, adapterRegistry, audit, requireScope, validateCsrf, requireFreshSession, confirmations }) {
  if (!confirmations?.issue || !confirmations?.consume) throw new Error('provider_confirmation_store_required');
  app.get('/api/v1/provider/protocols', async (request) => { await requireScope(request, 'provider.config.read'); return { items: adapterRegistry.list().map(({ protocolType, protocolVersion, descriptorVersion, taskModes, status }) => ({ protocolType, adapterVersion: protocolVersion, capabilityDescriptorSchemaVersion: descriptorVersion, configSchemaRef: `dgos.provider-config/${protocolType}/${protocolVersion}`, taskModes, status: status ?? 'active' })) }; });
  app.get('/api/v1/provider/capability-protocols', async (request) => { const auth = await requireScope(request, 'provider.config.read'); return directory.list(auth.subjectId); });
  app.post('/api/v1/provider/capability-protocols', async (request) => { await validateCsrf(request); const auth = await requireScope(request, 'provider.protocol.write'); await requireFreshSession(request, auth); session(auth); const input = body(request, ['requestId','declaration']); const declaration = validateTextProfile(input.declaration); return { ...declaration, digest: protocolDigest(declaration), status: 'validating' }; });
  app.post('/api/v1/provider/capability-protocols/confirmations', async (request, reply) => {
    await validateCsrf(request); const auth = await requireScope(request, 'provider.protocol.write'); await requireFreshSession(request, auth);
    const input = body(request, ['requestId','operation','protocolId','version','declaration','validationDigest','baseVersion','state']);
    if (!input.requestId || !input.protocolId || !input.version) fail('invalid_request');
    let digest;
    if (input.operation === 'provider.protocol.publish') { body(request, ['requestId','operation','protocolId','version','declaration','validationDigest']); if (typeof input.validationDigest !== 'string') fail('invalid_request'); const declaration = validateTextProfile(input.declaration); if (declaration.id !== input.protocolId || declaration.version !== input.version || protocolDigest(declaration) !== input.validationDigest) fail('version_conflict', 409); digest = input.validationDigest; }
    else if (input.operation === 'provider.protocol.state') { body(request, ['requestId','operation','protocolId','version','baseVersion','state']); if (!['active','disabled'].includes(input.state) || !Number.isInteger(input.baseVersion) || input.baseVersion < 1) fail('invalid_request'); digest = stateDigest(input); }
    else fail('invalid_request');
    return reply.code(201).send(await confirmations.issue({ subjectId: auth.subjectId, sessionId: session(auth), operation: input.operation, resourceId: input.protocolId, version: input.version, digest, requestId: input.requestId }));
  });
  app.post('/api/v1/provider/capability-protocols/:protocolId/versions', async (request, reply) => { await validateCsrf(request); const auth = await requireScope(request, 'provider.protocol.write'); await requireFreshSession(request, auth); const input = body(request, ['requestId','declaration','validationDigest','confirmationId']); const declaration = validateTextProfile(input.declaration); if (declaration.id !== request.params.protocolId || protocolDigest(declaration) !== input.validationDigest) fail('version_conflict', 409); const confirmation = { store: confirmations, input: { subjectId: auth.subjectId, sessionId: session(auth), operation: 'provider.protocol.publish', resourceId: declaration.id, version: declaration.version, digest: input.validationDigest, requestId: input.requestId, confirmationId: input.confirmationId } }; return reply.code(201).send(await directory.publish({ ownerId: auth.subjectId, protocolType: 'openai-compatible', profile: declaration, audit, requestId: input.requestId, confirmation })); });
  app.post('/api/v1/provider/capability-protocols/:protocolId/versions/:version/state', async (request) => { await validateCsrf(request); const auth = await requireScope(request, 'provider.protocol.write'); await requireFreshSession(request, auth); const input = body(request, ['requestId','baseVersion','state','confirmationId']); if (!['active','disabled'].includes(input.state) || !Number.isInteger(input.baseVersion) || input.baseVersion < 1) fail('invalid_request'); const confirmation = { store: confirmations, input: { subjectId: auth.subjectId, sessionId: session(auth), operation: 'provider.protocol.state', resourceId: request.params.protocolId, version: request.params.version, digest: stateDigest({ protocolId: request.params.protocolId, version: request.params.version, baseVersion: input.baseVersion, state: input.state }), requestId: input.requestId, confirmationId: input.confirmationId } }; return directory.setState({ ownerId: auth.subjectId, protocolId: request.params.protocolId, version: request.params.version, baseVersion: input.baseVersion, state: input.state, audit, requestId: input.requestId, confirmation }); });
}