# docs/04-技术架构/当前版本/V1-openapi.yaml
openapi: 3.0.3
info:
  title: DGOS V1 Logical API
  version: "1.0.0"
  description: |
    DGOS V1 逻辑契约的 HTTPS JSON + SSE 投影。桌面 IPC 必须保持相同的动作、字段、权限和事件语义。
    该文件同时登记已实现的本地 contract projection 与尚未具备真实桌面/Web E2E 证据的 draft operation。
security:
  - dgosSession: []
  - dgosApiKey: []
servers:
  - url: https://localhost
    description: Docker Compose Web projection (planned)
...
paths:
...
  /api/v1/system/settings:
    get:
      tags: [system]
      operationId: readSystemSettings
      x-tag-feature: V1-FR-001
      x-tag-implementation-status: implemented-local-contract
      responses:
        '200':
          description: Current system settings and effective values
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SystemSettingsSnapshot' }
    patch:
      tags: [system]
      operationId: writeSystemSettings
      x-tag-feature: V1-FR-001
      x-tag-implementation-status: implemented-local-contract
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SystemSettingsPatch' }
      responses:
        '200':
          description: Updated settings and restart status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SystemSettingsSnapshot' }
        '409':
          description: Settings version conflict
...
  /api/v1/provider/capability-protocols:
    get:
      tags: [providers]
      operationId: listCapabilityProtocols
      description: Authenticated provider.config.read; returns persisted active V1 text.chat protocol versions only. Read does not refresh models or enable policy.
      x-tag-feature: V1-FR-007
      x-tag-implementation-status: planned
      responses:
        '200':
          description: Registered declarative capability protocols
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/CapabilityProtocolRegistration' }
    post:
      tags: [providers]
      operationId: validateCapabilityProtocol
      security: [{ dgosSession: [] }]
      description: Validation only, no publish/enable/upstream request. Requires provider.protocol.write and a fresh authorized administrator/developer session; request fields validated against the V1 text-only declaration.
      x-tag-feature: V1-FR-007
      x-tag-implementation-status: planned
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CapabilityProtocolValidationRequest' }
      responses:
        '200':
          description: Canonical validated protocol
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CapabilityProtocolRegistration' }
        '422':
          description: Protocol schema or safety validation failed
  /api/v1/provider/capability-protocols/confirmations:
    post:
      tags: [providers]
      operationId: issueCapabilityProtocolConfirmation
      x-tag-feature: V1-FR-007
      x-tag-implementation-status: planned
      security: [{ dgosSession: [] }]
      description: Explicit review/confirm action; requires CSRF, provider.protocol.write, fresh administrator session and server-derived session ID. No API Key authentication. Issues a five-minute single-use ticket only; does not publish or change state. Same subject/requestId and identical session/operation/protocolId/version/payload digest return the original ticket without extending expiry; changed binding conflicts. Publication/state must reuse this requestId. Client echoes server validationDigest and never computes or supplies identity/digest for state changes.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CapabilityProtocolConfirmationRequest' }
      responses:
        '201':
          description: Issued or identically replayed ticket; no declaration/state mutation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CapabilityProtocolConfirmationReceipt' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/v1/provider/capability-protocols/{protocolId}/versions:
    parameters:
      - { in: path, name: protocolId, required: true, schema: { type: string } }
    post:
      operationId: registerCapabilityProtocolVersion
      security: [{ dgosSession: [] }]
      x-tag-feature: V1-FR-007
      x-tag-implementation-status: planned
      description: Controlled immutable publication. Requires CSRF, provider.protocol.write, fresh administrator Session and ticket from issueCapabilityProtocolConfirmation with the same requestId. Bind subject/session/operation/id/version/validated digest. An authorized identical completed request returns its stored result before checking ticket consumption/expiry; changed binding conflicts. For a new operation, commit declaration, operation result, audit/outbox and ticket consumption atomically; rollback consumes nothing. Never changes existing ProviderConfig or model policy.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CapabilityProtocolPublishRequest' }
      responses:
        '201':
          description: Persisted registered version
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CapabilityProtocolRegistration' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '428': { $ref: '#/components/responses/ConfirmationRequired' }
  /api/v1/provider/capability-protocols/{protocolId}/versions/{version}/state:
    parameters:
      - { in: path, name: protocolId, required: true, schema: { type: string } }
      - { in: path, name: version, required: true, schema: { type: string } }
    post:
      operationId: setCapabilityProtocolState
      security: [{ dgosSession: [] }]
      x-tag-feature: V1-FR-007
      x-tag-implementation-status: planned
      description: CSRF, provider.protocol.write and fresh administrator Session required. Use the issuance requestId and ticket bound to protocolId/version/baseVersion/state; baseVersion is registryVersion, not SemVer. Authorized identical completed replay returns its saved result without another increment/consumption; changed payload conflicts. New mutation commits state revision, operation result, audit/outbox and ticket consumption atomically. Disabling prevents new tasks, retains existing taskId/history, no automatic config migration.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CapabilityProtocolStateRequest' }
      responses:
        '200':
          description: Updated registry state, declaration bytes unchanged
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CapabilityProtocolRegistration' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '428': { $ref: '#/components/responses/ConfirmationRequired' }
...
components:
...
  securitySchemes:
    appLaunchTicket:
      type: apiKey
      in: query
      name: launchTicket
      description: Opaque short-lived session/subject/app/package/instance-bound static ticket, redacted from logs; no bridge authority.
    dgosSession:
      type: apiKey
      in: cookie
      name: dgos_session
      description: Revocable DGOS administrator or subject session cookie.
    dgosApiKey:
      type: http
      scheme: bearer
      bearerFormat: DGOS-API-Key
      description: DGOS API Key; the secret is never returned after creation.
  responses:
    ConfirmationRequired:
      description: confirmation_required; ticket missing, expired, consumed by another intent or mismatched. No protected mutation; fresh authentication failures use 403 step_up_required.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Unauthorized:
      description: Authentication missing, expired, or revoked
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Subject lacks the required capability or scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Conflict:
      description: Idempotency, version, reservation, or resource conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: Request failed schema or domain validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [errorKey, message, requestId]
      properties:
        errorKey: { type: string, description: Stable machine-readable error key }
        message: { type: string, description: Safe human-readable message without secrets }
        requestId: { type: string }
        retryable: { type: boolean }
        details: { type: object, additionalProperties: true, description: Redacted diagnostic details }
...
    AppearanceSettings:
      type: object
      required: [appearanceMode, windowMaterial, interfaceMode, displayScale, wallpaperFollowsAppearance]
      properties:
        appearanceMode: { type: string, enum: [light, dark, system] }
        windowMaterial: { type: string }
        interfaceMode: { type: string }
        displayScale: { type: number, enum: [0.75, 1.0, 1.25, 1.5, 1.75] }
        lightWallpaperRef: { type: string }
        darkWallpaperRef: { type: string }
        wallpaperFollowsAppearance: { type: boolean }
    LocaleSettings:
      type: object
      required: [uiLocale, effectiveLocale, regionFormat, assistantLanguage, projectContentLanguage]
      properties:
        uiLocale: { type: string }
        effectiveLocale: { type: string }
        regionFormat: { type: string }
        assistantLanguage: { type: string }
        projectContentLanguage: { type: string }
        fallbackState: { type: string, enum: [none, partial, full] }
    NetworkProxySettings:
      type: object
      required: [proxyMode, effectiveRoute, affectedServices, restartRequired]
      properties:
        proxyMode: { type: string, enum: [system, manual, off] }
        manualProxyRef: { type: string, description: Secret boundary reference; never a credential value }
        effectiveRoute: { type: string }
        affectedServices: { type: array, items: { type: string } }
        restartRequired: { type: boolean }
...
    GridSettings:
      type: object
      required: [enabled, style, spacing, majorLineEvery, showAxes, snapEnabled, snapTolerance, colorToken, opacity]
      properties:
        enabled: { type: boolean }
        style: { type: string, enum: [dot, line] }
        spacing: { type: number, minimum: 1 }
        majorLineEvery: { type: integer, minimum: 1 }
        showAxes: { type: boolean }
        snapEnabled: { type: boolean }
        snapTolerance: { type: number, minimum: 0 }
        colorToken: { type: string }
        opacity: { type: number, minimum: 0, maximum: 1 }
    PermissionRule:
      type: object
      required: [appId, subjectType, subjectId, capability, scope, decision]
      properties:
        appId: { type: string }
        subjectType: { type: string, enum: [user, organization, team, system] }
        subjectId: { type: string }
        capability: { type: string }
        scope: { type: object, additionalProperties: true }
        decision: { type: string, enum: [allow, ask, deny] }
        source: { type: string, enum: [manifest, user, admin, system] }
    SystemSettingsSnapshot:
      type: object
      additionalProperties: false
      required: [settingsVersion, appearance, locale, network, grid, restartRequired]
      properties:
        settingsVersion: { type: string }
        appearance: { $ref: '#/components/schemas/AppearanceSettings' }
        locale: { $ref: '#/components/schemas/LocaleSettings' }
        network: { $ref: '#/components/schemas/NetworkProxySettings' }
        grid: { $ref: '#/components/schemas/GridSettings' }
        privacy: { type: object, additionalProperties: true }
        appPermissions: { type: array, items: { $ref: '#/components/schemas/PermissionRule' } }
        restartRequired: { type: boolean }
        affectedServices: { type: array, items: { type: string } }
    SystemSettingsPatch:
      type: object
      additionalProperties: false
      required: [requestId, baseVersion, domain, patch]
      description: Unique HTTP projection. Domain fields only in patch; reject old nested patch.domain/value. Merge and validate the full domain under versioned transaction; no partial writes.
      properties:
        requestId: { type: string }
        baseVersion: { type: string }
        domain: { type: string, enum: [appearance, locale, network, grid, privacy, appPermissions] }
        patch: { type: object, additionalProperties: true, description: Only declared writable fields of selected domain, no derived effective-value override }
...
    CapabilityProtocolRegistration:
      type: object
      additionalProperties: false
      required: [schemaVersion, kind, id, version, executor, capabilities, operations, workflows, modelProfiles]
      properties:
        schemaVersion: { type: string, enum: [dgos-capability/v1] }
        kind: { type: string, enum: [model] }
        id: { type: string, pattern: '^[a-z0-9][a-z0-9._-]*$' }
        version: { type: string, pattern: '^(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)(?:-[0-9A-Za-z.-]+)?(?:\+[0-9A-Za-z.-]+)?$', description: Immutable exact declaration SemVer, not numeric registry revision }
        registryVersion: { type: integer, minimum: 1, readOnly: true, description: State concurrency revision incremented on every successful mutation, distinct from declaration version }
        digest: { type: string, readOnly: true }
        label: { type: string }
        summary: { type: string }
        executor:
          type: object
          additionalProperties: false
          required: [type, engine]
          properties:
            type: { type: string, enum: [declarative] }
            engine: { type: string, enum: [dgos-text-v1] }
        capabilities: { type: array, minItems: 1, maxItems: 1, items: { type: string, enum: [text.chat] } }
        operations: { type: object, minProperties: 1, additionalProperties: { $ref: '#/components/schemas/TextProtocolOperation' } }
        workflows:
          type: object
          additionalProperties: false
          required: [text.chat]
          properties:
            text.chat:
              type: object
              additionalProperties: false
              required: [submit]
              properties:
                submit: { type: string, description: Key of a declared text operation; no poll/download/upload or arbitrary workflow }
        modelProfiles: { type: object, minProperties: 1, additionalProperties: { $ref: '#/components/schemas/TextModelProfile' } }
        defaults: { $ref: '#/components/schemas/TextProfileDefaults' }
        limits: { $ref: '#/components/schemas/TextProfileLimits' }
        uiSchemas: { $ref: '#/components/schemas/TextProfileUi' }
        assets: { type: object, maxProperties: 0, additionalProperties: false }
        status: { type: string, enum: [draft, validating, active, rejected, disabled], readOnly: true }
        sourceRef: { type: string, readOnly: true, description: Controlled provenance assigned by registry, never a fetchable caller URL }
    TextProtocolOperation:
      type: object
      additionalProperties: false
      required: [profile, method, path]
      properties:
        profile: { type: string, enum: [chat.completions, responses] }
        method: { type: string, enum: [POST] }
        path: { type: string, enum: [/chat/completions, /responses], description: Must match profile; host/credentials/headers only from installed adapter and ProviderConfig }
    TextModelProfile:
      type: object
      additionalProperties: false
      required: [modelNames, workflow]
      properties:
        modelNames: { type: array, minItems: 1, uniqueItems: true, items: { type: string, minLength: 1 }, description: Exact model IDs or explicitly declared compatible aliases; no guessing or wildcard fallback }
        workflow: { type: string, enum: [text.chat] }
        defaults: { $ref: '#/components/schemas/TextProfileDefaults' }
        limits: { $ref: '#/components/schemas/TextProfileLimits' }
        uiSchemas: { $ref: '#/components/schemas/TextProfileUi' }
    TextProfileDefaults:
      $ref: '#/components/schemas/TextTaskParameters'
    TextTaskParameters:
      type: object
      additionalProperties: false
      description: Only text parameters recognized by the bound Adapter/model may be sent. Defaults merge by key; tighter descriptor/Profile/platform limits apply server-side. Unsupported parameter, non-finite number or bound violation rejects before Task/reservation/upstream side effects; no clamping/coercion.
      properties:
        temperature: { type: number, minimum: 0, maximum: 2 }
        maxOutputTokens: { type: integer, minimum: 1, maximum: 9007199254740991 }
    TextProfileLimits:
      type: object
      additionalProperties: false
      properties:
        maxInputCharacters: { type: integer, minimum: 1 }
        maxOutputTokens: { type: integer, minimum: 1 }
    TextProfileUi:
      type: object
      additionalProperties: false
      properties:
        parameters: { type: array, uniqueItems: true, items: { type: string, enum: [temperature, maxOutputTokens] }, description: Parameters must also be supported by the installed adapter/model descriptor; no script/HTML or arbitrary schema execution }
    CapabilityProtocolValidationRequest:
      type: object
      additionalProperties: false
      required: [requestId, declaration]
      properties:
        requestId: { type: string, format: uuid }
        declaration: { $ref: '#/components/schemas/CapabilityProtocolRegistration' }
    CapabilityProtocolPublishRequest:
      type: object
      additionalProperties: false
      required: [requestId, declaration, validationDigest, confirmationId]
      properties:
        requestId: { type: string, format: uuid }
        declaration: { $ref: '#/components/schemas/CapabilityProtocolRegistration' }
        validationDigest: { type: string, description: Recomputed by server; must bind identical validated declaration }
        confirmationId: { type: string, format: uuid, description: Ticket from issueCapabilityProtocolConfirmation; same requestId and full binding required }
    CapabilityProtocolConfirmationRequest:
      oneOf:
        - $ref: '#/components/schemas/CapabilityProtocolPublishConfirmation'
        - $ref: '#/components/schemas/CapabilityProtocolStateConfirmation'
      discriminator:
        propertyName: operation
        mapping:
          provider.protocol.publish: '#/components/schemas/CapabilityProtocolPublishConfirmation'
          provider.protocol.state: '#/components/schemas/CapabilityProtocolStateConfirmation'
    CapabilityProtocolPublishConfirmation:
      type: object
      additionalProperties: false
      required: [requestId, operation, protocolId, version, declaration, validationDigest]
      properties:
        requestId: { type: string, format: uuid }
        operation: { type: string, enum: [provider.protocol.publish] }
        protocolId: { type: string, description: Must equal declaration.id }
        version: { type: string, description: Must equal declaration.version; immutable exact SemVer }
        declaration: { $ref: '#/components/schemas/CapabilityProtocolRegistration' }
        validationDigest: { type: string, pattern: '^[a-f0-9]{64}$', description: Echo validation digest; server recomputes from validated declaration }
    CapabilityProtocolStateConfirmation:
      type: object
      additionalProperties: false
      required: [requestId, operation, protocolId, version, baseVersion, state]
      properties:
        requestId: { type: string, format: uuid }
        operation: { type: string, enum: [provider.protocol.state] }
        protocolId: { type: string, minLength: 1 }
        version: { type: string, description: Immutable exact declaration SemVer }
        baseVersion: { type: integer, minimum: 1 }
        state: { type: string, enum: [active, disabled] }
    CapabilityProtocolConfirmationReceipt:
      type: object
      additionalProperties: false
      required: [confirmationId, requestId, expiresAt, digest]
      properties:
        confirmationId: { type: string, format: uuid }
        requestId: { type: string, format: uuid }
        expiresAt: { type: string, format: date-time, description: Five minutes after first issuance; identical reissue never extends it }
        digest: { type: string, pattern: '^[a-f0-9]{64}$', description: Server-bound declaration digest or state payload digest; no subject/session/secret projection }
    CapabilityProtocolStateRequest:
      type: object
      additionalProperties: false
      required: [requestId, baseVersion, state, confirmationId]
      properties:
        requestId: { type: string, format: uuid }
        baseVersion: { type: integer, minimum: 1 }
        state: { type: string, enum: [active, disabled] }
        confirmationId: { type: string, format: uuid }
...